+++ src/Models/Queries/SectionQuery.php
@@ -18,7 +18,7 @@ class SectionQuery extends I18nActiveQuery
      */
     public function selectSiteAttributes(): static
     {
-        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->attributes(), [
+        return $this->addSelect($this->prefixColumns(array_diff($this->getModelInstance()->getColumnAttributes(), [
             'updated_by_user_id',
             'updated_at',
             'created_at',